[FVR-190] Puma - Add request queue depth
Closes: https://nitro.powerhrg.com/runway/backlog_items/FVR-190
Depends on: https://github.com/powerhome/nitro-web/pull/39227

Chart the Puma pod's Request Queue depth over time.

We want to measure the number of connections waiting to be accepted by
the Puma process. Previously, we'd assumed that the `puma_backlog`
measure this. However, "backlog" measures the number of client
connections which have already been accepted, but are awaiting
assignment to a Puma worker. Empirically, this number is always 0.

https://github.com/powerhome/nitro-web/pull/39227 introduces a custom
measure to track the request queue and we'd like to plot it in the
Puma dashboard.