Filter sum of memory request to only pods that are running (#74)
We added this filter to the CPU but for some reason we forgot about the
memory. With this the value should exclude pods that are left in the
namespace in a `Failed` or `Succeeded` state, normally the result of a
Job, since those pods don't have containers running, they are there just
to keep its metadata and logs until the job is deleted, they are not
considered for quota or the kube scheduler.